Move releases tab to its own file
